[maui] Experimenting with login flow
