+++ src/MvpCompanion.Maui/ShellPage.xaml
@@ -6,7 +6,7 @@
        xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
        ios:Page.UseSafeArea="True"
        FlyoutBehavior="{OnIdiom Phone=Disabled, Default=Locked}"
-       FlyoutWidth="68">
+       FlyoutWidth="50">
 
     <!-- Desktop/Tablet-->
     <FlyoutItem Title="Home"
@@ -45,16 +45,6 @@
             <ShellContent ContentTemplate="{DataTemplate views:Settings}" />
         </Tab>
     </TabBar>
-
-    <primitives:RadPopup.Popup>
-        <primitives:RadPopup x:Name="LoginPopup"
-                             Placement="Center"
-                             IsModal="True"
-                             OutsideBackgroundColor="#6F000000">
-            <WebView x:Name="_webView"
-                     Navigated="WebView_OnNavigated" />
-        </primitives:RadPopup>
-    </primitives:RadPopup.Popup>
 
     <Shell.FlyoutFooterTemplate>
         <DataTemplate>
