+++ src/MvpCompanion.Maui/Views/Home.xaml
@@ -6,6 +6,19 @@
              xmlns:primitives="clr-namespace:Telerik.XamarinForms.Primitives;assembly=Telerik.Maui.Controls.Compatibility"
              Title="Home">
     <Grid>
+        <CollectionView ItemsSource="{Binding Contributions}">
+            <CollectionView.ItemTemplate>
+                <DataTemplate>
+                    <StackLayout>
+                        <Label Text="{Binding StartDate}" />
+                        <Label Text="{Binding Title}" />
+                        <Label Text="{Binding ContributionTypeName}" />
+                        <Label Text="{Binding ReferenceUrl}" />
+                    </StackLayout>
+                </DataTemplate>
+            </CollectionView.ItemTemplate>
+        </CollectionView>
+
         <primitives:RadBusyIndicator x:Name="BusyIndicator"
                                      IsVisible="{Binding IsBusy}"
                                      IsBusy="{Binding IsBusy}"
